Simplistic ESS simulation within a intraspecies-symmetric population
This is the first version of what will be the simulation of interaction
between subjects within a population of the same species.
For now, the subjects have only one characteristic that affects how well
they'll do in the gene pool: aggressive behavior.
Each subject has a <geneticAggressiveness>, and will TEND to predict
how the subjects behaves.
This <geneticAggressiveness> is inheritable to the childs by taking the
average of this value for the parents.

Sex is coded into the subjects, but (for now) it doesn't have any
effects in the gene pool.

The success inside the gene pool is measured by "points", calculated by
simulating fights between members of the population. Losing, winning,
being hurt, and losing time, affect these points.

Every generation, all those subjects that are both a) above 0 points, and
b) above the average score, get the right to reproduce for the next
generation with any other successful subject. If those two conditions are
met, it doesn't matter if subject A has 10000 points, and subject B has 5,
they're considered equally successful.

--- Notes for next version ---
- Take sex into account
- Make points  more relevant in the quantity of offspring
- Refactor the code
- Make fighting a more complex task (if you're recently injured,
  you'll most likely lose, for example)
- Add more behaviors to the population (retaliators, and bullies)
- Make "fight-aptness" inheritable